+++ lib/drag_list.dart
@@ -1,3 +1,3 @@
 library drag_list;
 
-export 'package:drag_list/src/DragList.dart';
+export 'package:drag_list/src/drag_list.dart';
