+++ lib/src/axis_dimen.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
 
 mixin AxisDimen {
   Axis get axis;
